+++ src/services/chat/chat.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { URL } from './client';
+// import { URL } from '../client';
 import { Client, Frame, Message } from '@stomp/stompjs';
 import { io } from 'socket.io-client';
