--- src/services/client.ts
@@ -1,1 +1,0 @@
-export const URL = 'http://34.172.123.179';
