+++ src/services/pin/putPinEdit.ts
@@ -1,4 +1,4 @@
-import { api } from '.';
+import { api } from '..';
 
 export interface PinEditData {
   boardNo: number | null;
